+++ tests/testthat/test-get_predictors.R
@@ -2,10 +2,10 @@ library(testthat)
 
 test_that("get_predictors returns predictor names for glmnet model", {
     set.seed(1)
-    frm <- train_frm(df = RP[1:40, ], method = "lasso", nfolds = 2, nw = 1, verbose = 0)
+    frm <- train_frm(df = RP[1:40, ], method = "lasso", nfolds = 2, nw = 1, verbose = 2, do_cv = FALSE)
     preds <- get_predictors(frm)
     expect_true(is.character(preds))
-    expect_true(length(preds) > 10) # should have many descriptors
+    expect_true(length(preds) > 10)
 })
 
 test_that("get_predictors returns predictor names for gbtree model", {